refactored to work with a mongo db and activeMQ queue
