+++ AuditConsumer/.gitignore
@@ -29,3 +29,4 @@ build/
 
 ### VS Code ###
 .vscode/
+/.metadata/
