+++ AuditConsumer/Servers/Pivotal tc Server Developer Edition v4.0-config/context.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Context>
+    <!-- Default set of monitored resources -->
+    <WatchedResource>WEB-INF/web.xml</WatchedResource>
+    <!-- Uncomment this to disable session persistence across Tomcat restarts -->
+    <!--
+            <Manager pathname="" />
+            -->
+    <!-- Uncomment this to enable Comet connection tacking (provides events
+                 on session expiration as well as webapp lifecycle) -->
+    <!--
+            <Valve className="org.apache.catalina.valves.CometConnectionManagerValve" />
+            -->
+    <WatchedResource>${catalina.base}/conf/web.xml</WatchedResource>
+</Context>
