+++ AuditConsumer/src/main/java/com/bae/repository/AuditRepository.java
@@ -1,9 +1,11 @@
 package com.bae.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
 import com.bae.domain.SentRequest;
 
-public interface AuditRepository extends MongoRepository<SentRequest, String>{
+@Repository
+public interface AuditRepository extends MongoRepository<SentRequest, String> {
 
 }
