--- AuditConsumer/src/main/java/com/bae/service/AuditListener.java
@@ -1,8 +1,0 @@
-package com.bae.service;
-
-import com.bae.domain.SentRequest;
-
-public interface AuditListener {
-
-	public void receiveMessage(SentRequest sentRequest);
-}
