+++ AuditConsumer/src/main/resources/application.properties
@@ -2,6 +2,7 @@ spring.activemq.user=admin
 spring.activemq.password=admin
 spring.activemq.broker-url=tcp://localhost:61616?jms.redeliveryPolicy.maximumRedeliveries=1
 activemq.queue.name=AuditQueue
+
 server.port=8008
 
 spring.data.mongodb.host=localhost
